# docs/Swagger.yaml
swagger: '2.0'
info:
  description: |
    AND1 Project
  version: pre-alpha-1.0
  title: Event & Equipment Manager
paths:
  /token:
    get:
      summary: returns jwt, provided that the username and password is correct
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Basic Authorization
      produces:
      - text/plain
      responses:
        200:
          description: JSON Web Token using HS256 algorithm
          schema:
            example: Header.Payload.Signature
        401:
          description: Unauthorized
          schema:
            example: Incorrect credentials
  /user:
    get:
      summary: returns all users
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      produces:
      - application/json
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/user'
        401:
          description: Unauthorized
          schema:
            example: Authorization failed!
        403:
          description: Forbidden
          schema:
            example: You don't have high enough security clearance for this operation!
    post:
      summary: creates a new user
      parameters:
      - in: header
        name: Authorization
        required: false
        type: string
        description: Bearer Authorization
      - in: body
        name: user
        description: user to create
        required: true
        schema:
            $ref: '#/definitions/user'
      produces:
      - application/json
      responses:
        200:
          description: Created
        401:
          description: Unauthorized
          schema:
            example: Authorization failed!
        403:
          description: Forbidden
          schema:
            example: You don't have high enough security clearance for this operation!
        409:
          description: Conflict
          schema:
            example: username taken
  /user/id/{id}:
    get:
      summary: returns the user with the given id
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: id
        in: path
        description: id of the user to return
        required: true
        type: integer
      produces:
      - application/json
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/user'
        401:
          description: Unauthorized
          schema:
            example: Authorization failed!
        403:
          description: Forbidden
          schema:
            example: You don't have high enough security clearance for this operation!
        404:
          description: Not Found
    delete:
      summary: deletes the user with the given id
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: id
        in: path
        description: id of the user to delete
        required: true
        type: integer
      produces:
      - application/json
      responses:
        200:
          description: Deleted
        401:
          description: Unauthorized
          schema:
            example: Authorization failed!
        403:
          description: Forbidden
          schema:
            example: You don't have high enough security clearance for this operation!
        404:
          description: Not Found
    put:
      summary: updates the user with the given id
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: id
        in: path
        description: id of the user to update
        required: true
        type: integer
      - in: body
        name: user
        description: user details to update
        required: true
        schema:
            $ref: '#/definitions/user'
      produces:
      - application/json
      responses:
        200:
          description: Updated
        401:
          description: Unauthorized
          schema:
            example: Authorization failed!
        403:
          description: Forbidden
          schema:
            example: You don't have high enough security clearance for this operation!
        404:
          description: Not Found
        409:
          description: Conflict
          schema:
            example: username taken
  /user/name/{username}:
    get:
      summary: returns the user with the given username
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: username
        in: path
        description: username of the user to return
        required: true
        type: string
      produces:
      - application/json
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/user'
        401:
          description: Unauthorized
          schema:
            example: Authorization failed!
        403:
          description: Forbidden
          schema:
            example: You don't have high enough security clearance for this operation!
        404:
          description: Not Found
  /user/me:
    get:
      summary: returns the authenticated user
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      produces:
      - application/json
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/user'
        401:
          description: Unauthorized
          schema:
            example: Authorization failed!
...
definitions:
...
  user:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: Bob
      password:
        type: string
        example: Kolbasz1234
      permission:
        type: string
        example: employee
      organization:
        type: string
        example: "Event-Audio limited company"